% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_trend_seasonal_model}
\alias{fit_trend_seasonal_model}
\title{Fit trend + seasonal model using ridge regression (glmnet, alpha = 0)}
\usage{
fit_trend_seasonal_model(y, seasonal_periods, n_changepoints)
}
\description{
Fit trend + seasonal model using ridge regression (glmnet, alpha = 0)
}
